Break dependencies. This is why kitchen sink utils modules suck.
